Fixed cart quantity issue

+++ actions/customer/ProductAndStore/Cart.Action.ts
@@ -288,7 +288,7 @@ export const RemoveItem = async (
     cart.items.splice(index, 1);
   }
   await cart.save();
-  if (cart.items.length === 0 && cart.subsidyItems.length === 0) {
+  if (cart.items.length === 0 && cart.subsidyItems.length === 0 && cart.miscItems.length === 0) {
     await CartModel.findOneAndDelete({ customerId: user._id });
   }
   revalidatePath("/customer/cart");
